+++ R/broom.helpers-package.R
@@ -12,13 +12,31 @@ dplyr::`%>%`
 # cf. https://github.com/r-lib/tidyselect/issues/201
 utils::globalVariables("where")
 
+# Escapes any characters that would have special
+# meaning in a regular expression
+# based on Hmisc::escapeRegex
+.escapeRegex <- function (string)
+{
+  gsub("([.|()\\^{}+$*?]|\\[|\\])", "\\\\\\1",
+       string)
+}
+
+
 # remove backtips around variable names
 .clean_backtips <- function(x, variable_names = x) {
   for (i in stats::na.omit(variable_names)) {
-    x <- stringr::str_replace_all(x, paste0("`", i, "`"), i)
+    x <- stringr::str_replace_all(
+      x,
+      paste0("`", .escape_regex(i), "`"),
+      i
+    )
 
     if (stringr::str_detect(i, "^`.*`$")) {
-      x <- stringr::str_replace_all(x, i, stringr::str_sub(i, 2, -2))
+      x <- stringr::str_replace_all(
+        x,
+        .escape_regex(i),
+        stringr::str_sub(i, 2, -2)
+      )
     }
   }
   x