+++ NEWS.md
@@ -1,5 +1,9 @@
 # broom.helpers (development version)
 
+* Bug fix for non standard variable names containing
+  a character that would have a special meaning in
+  a regular expression (#44)
+
 # broom.helpers 1.0.0
 
 * Initial version